处理订单详情中商品点击跳到商品去

+++ app/src/main/java/com/zhaidou/fragments/OrderUnPayFragment.java
@@ -205,7 +205,7 @@ public class OrderUnPayFragment extends BaseFragment {
                             JSONObject orderObj = orderArr.optJSONObject(i);
                             int id = orderObj.optInt("id");
                             String number = orderObj.optString("number");
-                            double amount = orderObj.optInt("amount");
+                            double amount = orderObj.optDouble("amount");
                             String status = orderObj.optString("status");
                             String status_ch = orderObj.optString("status_ch");
                             String created_at = orderObj.optString("created_at");
